+++ src/models/persistence/DesiredShutterPosition.ts
@@ -1,0 +1,3 @@
+export class DesiredShutterPosition {
+  constructor(public desiredPosition: number) {}
+}